+++ megaparse/__init__.py
@@ -1,3 +1,8 @@
-# from .Converter import MegaParse
+"""My library with optional components."""
 
-# __all__ = ["MegaParse"]
+__version__ = "0.1.0"
+
+# Import only SDK components by default
+from megaparse import sdk
+
+__all__ = ["sdk"]
