+++ megaparse/sdk/__init__.py
@@ -1,0 +1,3 @@
+from megaparse.sdk.src import MegaParseSDK
+
+__all__ = ["MegaParseSDK"]
