+++ megaparse/sdk/src/__init__.py
@@ -4,7 +4,7 @@ from .endpoints.url_upload import URLUpload
 
 
 class MegaParseSDK:
-    def __init__(self, base_url: str, api_key: str):
+    def __init__(self, api_key: str):
         self.client = MegaParseClient(api_key)
         self.file = FileUpload(self.client)
         self.url = URLUpload(self.client)
