+++ megaparse/sdk/src/endpoints/url_upload.py
@@ -1,14 +1,10 @@
-from .client import MegaParseClient
-import asyncio
+from megaparse.sdk.src.client import MegaParseClient
 
 
 class URLUpload:
     def __init__(self, client: MegaParseClient):
         self.client = client
 
-    async def aupload(self, url: str):
+    async def upload(self, url: str):
         data = {"url": url}
-        return self.client.request("POST", "/url", json=data)
-
-    def upload(self, url: str):
-        return asyncio.run(self.aupload(url))
+        return await self.client.request("POST", "/url", json=data)
