+++ pyproject.toml
@@ -8,25 +8,19 @@ authors = [
 ]
 
 readme = "README.md"
+requires-python = ">= 3.11"
+
 dependencies = [
-    "python-docx>=1.1.0",
     "mammoth>=1.8.0",
-    "python-pptx>=1.0.2",
-    "llama-parse>=0.4.0",
-    "pdf2docx>=0.5.0",
-    "unstructured[all-docs]>=0.15.0",
-    "langchain>=0.2.0",
-    "langchain-community>=0.2.0",
-    "langchain-openai>=0.1.0",
-    "langchain-core>=0.2.0",
     "python-dotenv>=1.0.0",
+
+]
+
+[project.optional-dependencies]
+core = [
     "pycryptodome>=3.21.0",
     "llama-index>=0.10.0",
     "pdfplumber>=0.11.0",
-    "fastapi>=0.115.2",
-    "uvicorn>=0.32.0",
-    "ratelimit>=2.2.1",
-    "requests>=2.32.3",
     "backoff>=2.2.1",
     "pypdf>=5.0.1",
     "psutil>=6.1.0",
@@ -34,10 +28,24 @@ dependencies = [
     "playwright>=1.47.0",
     "langchain-anthropic>=0.2.3",
     "python-magic>=0.4.27",
+    "unstructured[all-docs]>=0.15.0",
+    "langchain>=0.2.0",
+    "langchain-community>=0.2.0",
+    "langchain-openai>=0.1.0",
+    "langchain-core>=0.2.0",
+    "llama-parse>=0.4.0",
 ]
-requires-python = ">= 3.11"
+api = [
+    "uvicorn>=0.32.0",
+    "fastapi>=0.115.2",
+    "ratelimit>=2.2.1",
+    "requests>=2.32.3",
+]
 
-
+all = [
+    "megaparse[core]"
+    "megaparse[api]
+]
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
